Preview of experiments. FIXED bugs for show.php. NEW Pop-up before delete experiment

+++ mod/reservations/experiments/index.php
@@ -16,7 +16,7 @@ $lab = Laboratory::find_by_id($_GET["laboratory_id"]);
 if (!$lab){
   echo "Laboratorio Inv&aacute;lido";
   echo $OUTPUT->footer();
-  die();
+  exit();
 }
 ?>
 
@@ -29,6 +29,7 @@ if (!$lab){
       <th>&nbsp;</th>
       <th>&nbsp;</th>
       <th>&nbsp;</th>
+      <th>&nbsp;</th>
     </tr>
     <?php
   $experiments = $lab->experiments();
@@ -37,11 +38,12 @@ if (!$lab){
     echo "<td>" . $experiment->name . "</td>";
     echo "<td><a href='../contents/index.php?experiment_id=" . $experiment->id . "'>Ver Documentos</a></td>";
     if (has_capability("mod/reservations:update_experiment", $context)) {
-      echo "<td>". $experiment->activation_link() . "</td>";
-      echo "<td><a href='edit.php?experiment_id=" . $experiment->id . "'><img title='Editar' alt='Editar' src='".$OUTPUT->pix_url('f/edit')."'/></a></td>";
+      echo "<td>". $experiment->activation_link($OUTPUT->pix_url('i/show'), $OUTPUT->pix_url('i/hide')) . "</td>";
+      echo "<td><a href='../show.php?id=0&exp=" . $experiment->id . "'><img title='Vista Previa' alt='Vista Previa' src='" . $OUTPUT->pix_url('f/edit') . "'/></a></td>";
+      echo "<td><a href='edit.php?experiment_id=" . $experiment->id . "'><img title='Editar' alt='Editar' src='" . $OUTPUT->pix_url('a/setting') . "'/></a></td>";
     }
     if (has_capability("mod/reservations:delete_experiment", $context)) {
-      echo "<td><a href='delete.php?experiment_id=" . $experiment->id . "'><img title='Eliminar' alt='Eliminar' src='".$OUTPUT->pix_url('i/cross_red_big')."'/></a></td>";
+      echo "<td><a href='delete.php?experiment_id=" . $experiment->id . "' onclick='return confirm(\"Una vez borrado el experimento no se puede recuperar. Seguro que desea eliminar el experimento?\")'><img title='Eliminar' alt='Eliminar' src='" . $OUTPUT->pix_url('i/cross_red_big') . "'/></a></td>";
     }
     echo "</tr>";
   }
